requirements.txt
Flask
gunicorn
flask_sqlalchemy
flask_migrate
psycopg2-binary
python-dotenv
pytest
flake8